Remove `PipelineRewriterTest` and consolidate `$and` query tests into `IntersectionNodeWithIndexScanStrategyTest`, refactor `UpdateExecutor` to handle index value type mismatches, and improve `UnionNode` and related classes with enhanced documentation and query execution behavior. Add `ZGetF64HandlerTest` for testing float64 operations in ZMap and include minor cleanups across test classes.

+++ kronotop/src/main/java/com/kronotop/server/Response.java
@@ -67,7 +67,7 @@ public interface Response {
      *
      * @param value the double value to be written
      */
-    void writeDouble(long value);
+    void writeDouble(double value);
 
     /**
      * Writes an array of Redis messages as a response message to the client.
